Added sysctl.conf.new
Okay, so we'll make yet another way of tweaking conf files. The .new
file method familiar from SlackBuilds. This is only for files that are
expected to be new to the system.

We'll start with sysctl.conf for several reasons:

- It requires tweaking every now and then (easier to modify the .new vs.
  the patch file)
- It controls so many aspects of the system
- For easier referencing of it
- Newer version of procps now support this preload path thingie, that
  loads sysctl confs from /etc/sysctl.d. We'll probably utilize that
  when it lands to Slackware (see https://gitorious.org/procps/procps/commit/10095927)
- We don't have to duplicate it to quick_harden()

The old way of creating sysctl.conf from the etc patch will stay for now
(for backwards compatibility), but we'll switch over at some point.