Added create_limited_ca_list() to -A
Now the import_pgp_keys() is a bit safer, as wgets are run with limited CAs.

I think we still need to get rid of the whole wgetting of PGP keys,
because it's dangerous to do "gpg --import -" as an adversary might have
compromised the WWW servers and gpg will happily import these keys.

When gpg is used with --recv-keys, there are safeguards to check whether
the key actually is the one we requested and we can refuse it, even
though it would come from a hostile key server.