Configure default_ccache_name in krb5.conf to store tickets under /run/user/
This is because of pam_namespace, as rpc.gssd (running as root) can't
find user's tickets in /tmp/krb5cc_%{uid}, because they really exist in
/tmp-inst/.

rpc.gssd has a default search path that includes "/run/user/%U".

This is also better for security, as the Kerberos tickets are not laying
around in /tmp indefinitely, as /run/user/%{uid} gets removed after the
last session ends.