nftables: Set FORWARD chain policies as inet->accept and ip&ip6->drop
This is because of Docker, so when it uses the legacy iptables
interface, the forwarding rules go to "ip" table -> FORWARD chain and
the "inet" table's policy doesn't prevent it from working.

Because the "inet" table's policy seems to override what's in "ip" and
your containers are unable to communicate to the internet.