rbash: Use a separate pam_env entry and /etc/restricted_env
We wrap it around pam_succeed_if, so that it is only applied to users
who have /bin/rbash as shell. This is because it seems to mess up stuff
if we just put it into regular /etc/environment as the PATH contains
only /usr/local/rbin. At least Ansible was breaking up, as it couldn't
find commands.

Do note, that it currently only applies to sshd.

Without this a restricted user could just run "ssh user@host bash" and
call it a day. The PATH needs to be enforced early on.

More details in the wiki:
https://github.com/pyllyukko/harden.yml/wiki/Restricted-shell